refactor: update test gateways with offsite as well
