+++ src/PaymentGateways/Gateways/TestGateway/TestGateway.php
@@ -2,18 +2,14 @@
 
 namespace Give\PaymentGateways\Gateways\TestGateway;
 
-use Exception;
 use Give\Donations\Models\Donation;
 use Give\Framework\PaymentGateways\Commands\GatewayCommand;
 use Give\Framework\PaymentGateways\Commands\PaymentComplete;
 use Give\Framework\PaymentGateways\Commands\PaymentRefunded;
-use Give\Framework\PaymentGateways\Commands\SubscriptionComplete;
 use Give\Framework\PaymentGateways\PaymentGateway;
 use Give\Framework\Support\Facades\Scripts\ScriptAsset;
 use Give\Helpers\Form\Utils as FormUtils;
 use Give\PaymentGateways\Gateways\TestGateway\Views\LegacyFormFieldMarkup;
-use Give\Subscriptions\Models\Subscription;
-use Give\Subscriptions\ValueObjects\SubscriptionStatus;
 
 /**
  * Class TestGateway
@@ -93,34 +89,6 @@ class TestGateway extends PaymentGateway
         $intent = $gatewayData['testGatewayIntent'] ?? 'test-gateway-intent';
         
         return new PaymentComplete("test-gateway-transaction-id-{$intent}-$donation->id");
-    }
-
-    /**
-     * @inheritDoc
-     *
-     * @since 2.23.0
-     */
-    public function createSubscription(
-        Donation $donation,
-        Subscription $subscription,
-        $gatewayData
-    ): GatewayCommand {
-        return new SubscriptionComplete(
-            "test-gateway-transaction-id-$donation->id",
-            "test-gateway-subscription-id-$subscription->id"
-        );
-    }
-
-    /**
-     * @since 2.23.0
-     *
-     * @inheritDoc
-     * @throws Exception
-     */
-    public function cancelSubscription(Subscription $subscription)
-    {
-        $subscription->status = SubscriptionStatus::CANCELLED();
-        $subscription->save();
     }
 
     /**
