--- src/PaymentGateways/Gateways/TestOffsiteGateway/testOffsiteGateway.js
@@ -1,11 +1,0 @@
-(() => {
-    'use strict';
-    const testOffsiteGateway = {
-        id: 'test-gateway-offsite',
-        Fields() {
-            return givewpTestGatewayOffsiteData.message;
-        },
-    };
-
-    window.givewp.gateways.register(testOffsiteGateway);
-})();
